debug

+++ app/Services/ShipRocketService.php
@@ -42,21 +42,27 @@ class ShipRocketService
             // dd($response);
             log::info('status code for create order' . $response['status_code']);
             // $response = json_decode($response);
-            if ($response['status_code'] == 1) {
+            if (isset($response) && (!empty($response))) {
+                if ($response['status_code'] == 1) {
 
-                CartShiprocketResponse::where('cart_token', $params['order_id'])->delete();
-                $ins_params['cart_token'] = $params['order_id'];
-                $ins_params['rocket_token'] = $token;
-                $ins_params['request_type'] = 'create_order';
-                $ins_params['rocket_order_request_data'] = json_encode($params);
-                $ins_params['rocket_order_response_data'] = $response;
-                $ins_params['order_id'] = $response['order_id'];
-                $ins_params['brand_id'] = $brand_id;
+                    CartShiprocketResponse::where('cart_token', $params['order_id'])->delete();
+                    $ins_params['cart_token'] = $params['order_id'];
+                    $ins_params['rocket_token'] = $token;
+                    $ins_params['request_type'] = 'create_order';
+                    $ins_params['rocket_order_request_data'] = json_encode($params);
+                    $ins_params['rocket_order_response_data'] = $response;
+                    $ins_params['order_id'] = $response['order_id'];
+                    $ins_params['brand_id'] = $brand_id;
 
-                CartShiprocketResponse::create($ins_params);
+                    CartShiprocketResponse::create($ins_params);
+                } else {
+                    log::debug($response);
+                }
             } else {
-                log::debug($response);
+                log::debug($e);
+                return null;
             }
+
 
             return $response;
         } catch (Exception  $e) {
